switch from int data to double data, needed for neural network ops

+++ lib/matrix-math.h
@@ -11,7 +11,7 @@
 
 typedef struct {
   int rows, columns;
-  int **data;
+  double **data;
 } Matrix;
 
 /*
